done some more functions

+++ ft_substr.c
@@ -6,13 +6,13 @@
 /*   By: jarregui <jarregui@student.42madrid.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/07/14 13:00:53 by jarregui          #+#    #+#             */
-/*   Updated: 2021/07/14 13:02:56 by jarregui         ###   ########.fr       */
+/*   Updated: 2021/07/14 14:45:06 by jarregui         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-char *ft_substr(char const *s, unsigned int start, size_t len)
+char *ft_substr(char const *s, unsigned int start, unsigned int  len)
 {
 
 }
